Create README.md
1) Chemical composition of pottery
An unsupervised learning approach is used to investigate the multivariate chemical composition of the raw material from four different sources using Principal Component Analysis.
2) Batch Data Analysis
Batch manufacturing processes are very common in chemical, pharma, bioengineering and semi-conductor industries such as baker's yeast production, beer brewing and vaccines production. In real life, a typical batch is run from a few hours up to a week or two and a lot of things can go wrong during this period. There is always variability either because the process is very sensitive to minor fluctuations in some variables or the control of some variables failed for a period of time.
In a company that implements Data Analytics or Multivariate Statistical Process Control (MSPC) monitoring is typically implemented by creating a PCA model of the perfect batches identified.
3) PAT application:
PCA and PLS, two classic MVDA approaches are used to assess the feasibility of using NIR to measure a critical quality component of a wood product, the composition of wood fiber.